requirements.txt
asgiref==3.7.2
Brotli==1.1.0
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==4.2.7
django-appconf==1.0.6
django-googledrive-storage==1.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
google-api-core==2.10.2
google-api-python-client==2.108.0
google-auth==1.35.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
gunicorn==21.2.0
httplib2==0.22.0
httplib2shim==0.0.3
idna==3.6
inflection==0.5.1
oauth2client==4.1.3
packaging==23.2
Pillow==10.1.0
protobuf==4.25.1
psycopg==3.1.12
psycopg-binary==3.1.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
